+++ app/models/alchemy/content.rb
@@ -219,10 +219,7 @@ module Alchemy
 
     # Returns true if this content should be taken for element preview.
     def preview_content?
-      if definition['take_me_for_preview']
-        ActiveSupport::Deprecation.warn("Content definition's `take_me_for_preview` key is deprecated. Please use `as_element_title` instead.")
-      end
-      !!definition['take_me_for_preview'] || !!definition['as_element_title']
+      !!definition['as_element_title']
     end
 
     # Proxy method that returns the preview text from essence.